isolate collectSubjectMetadata
